Rename auth.ts for auth.lib.ts
