+++ src/app/api/route.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 import { type NextRequest } from 'next/server'
 
 import { getCookieProps } from "@/shared/utils/parseCookie";
-import { encodeAccessToken, saveSessionCookie } from "@/shared/lib/auth";
+import { encodeAccessToken, saveSessionCookie } from "@/shared/lib/auth.lib";
 import { GeneralError } from "@/shared/types/Global";
 
 export async function POST(request: NextRequest) {
