+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { LoginPage } from '@/features/Login/Login/LoginPage';
 import { DASHBOARD_ROUTE } from '@/shared/constants/Global.constants';
-import { getAccessToken } from '@/shared/lib/auth';
+import { getAccessToken } from '@/shared/lib/auth.lib';
 import { redirect } from 'next/navigation';
 
 export default async function Home() {
