+++ src/shared/lib/auth.lib.ts
